BIGDATA/perfume/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bleach==6.0.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
dask==2023.3.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-pandas==0.6.6
django-redis==5.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
executing==1.2.0
fsspec==2023.3.0
gunicorn==20.1.0
importlib-metadata==6.1.0
inflection==0.5.1
ipykernel==6.22.0
ipython==8.11.0
ipywidgets==8.0.6
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
locket==1.0.0
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
partd==1.3.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.5
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2022.7
uritemplate==4.1.1
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==4.0.7
zipp==3.15.0
